对原来服务代码重新整理，添加延时1s操作，对未抢到红包用户跳转红包详情
