+++ app/src/main/java/wallet/mysiga/com/redwallet/view/IWalletServiceView.java
@@ -1,0 +1,12 @@
+package wallet.mysiga.com.redwallet.view;
+
+import android.accessibilityservice.AccessibilityService;
+
+/**
+ * 服务接口
+ *
+ * @author Wlison milin411@163.com
+ */
+public interface IWalletServiceView {
+    AccessibilityService getAccessibilityService();
+}
