+++ app/src/main/res/layout/activity_main.xml
@@ -7,7 +7,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context=".MainActivity">
+    tools:context=".RedWalletConfigConfigActivity">
 
     <include layout="@layout/include_common_appbar" />
